Almost Completed Favorites

+++ nodbproject/src/components/ListContainer.js
@@ -2,10 +2,9 @@ import React, { Component } from 'react';
 import Playlists from './Playlists';
 import './CSS/listcontainer.css';
 
-export default function ListContainer(props) {
-    const { favorite,array,fav,deleteItem } = props;
+export default function ListContainer({ favorite,array,fav,deleteItem }) {
      
-    let playlists = props.array.map((cur,ind) => {
+    let playlists = array.map((cur,ind) => {
         if(cur.platformId === 1){
             return <Playlists key={ind} name={cur.name} platformName={'XBOX One'} platformId={cur.platformId}  
             population={cur.population.players} id={cur.id} favorite={favorite} fav={fav} deleteItem={deleteItem} /> 
@@ -14,7 +13,7 @@ export default function ListContainer(props) {
             return <Playlists key={ind} name={cur.name} platformName={'Play Station 4'} platformId={cur.platformId} 
             population={cur.population.players} id={cur.id} favorite={favorite} fav={fav} deleteItem={deleteItem} /> 
         }
-        else{
+        else if(cur.platformId === 3){
             return <Playlists key={ind} name={cur.name} platformName={'Steam'} platformId={cur.platformId} 
             population={cur.population.players} id={cur.id} favorite={favorite} fav={fav} deleteItem={deleteItem} /> 
         }
